pyproject.toml
[project]
name = "researchlab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "alirezanobakht13", email = "a.nobakht13@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "coolname>=3.0.0",
    "gitpython>=3.1.46",
    "mlflow>=3.9.0",
    "pyyaml>=6.0.3",
    "typer>=0.23.1",
]

[project.scripts]
rlab = "researchlab.tracking.cli:app"

[project.optional-dependencies]
design = [
    "chex>=0.1.91",
    "equinox>=0.13.4",
    "jax[cuda13]>=0.9.0.1",
    "jaxtyping>=0.3.9",
    "optax>=0.2.7",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
